Moved code to access fire to a common submodule
