+++ cmd-encryption/src/main/java/uk/ac/ebi/ega/cmd/encryption/services/FileReEncryptService.java
@@ -20,8 +20,8 @@ package uk.ac.ebi.ega.cmd.encryption.services;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.ac.ebi.ega.cmd.encryption.options.OutputFormat;
-import uk.ac.ebi.ega.cmd.encryption.services.fire.FireService;
-import uk.ac.ebi.ega.cmd.encryption.services.fire.IFireFile;
+import uk.ac.ebi.ega.fire.FireService;
+import uk.ac.ebi.ega.fire.IFireFile;
 import uk.ac.ebi.ega.encryption.core.ReEncryption;
 import uk.ac.ebi.ega.encryption.core.encryption.AesCbcOpenSSL;
 import uk.ac.ebi.ega.encryption.core.encryption.AesCtr256Ega;
