+++ file-re-encrypt/pom.xml
@@ -33,6 +33,12 @@
 
         <dependency>
             <groupId>uk.ac.ebi.ega</groupId>
+            <artifactId>fire-core</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>uk.ac.ebi.ega</groupId>
             <artifactId>file-encryption-core</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
