--- file-re-encrypt/src/main/java/uk/ac/ebi/ega/file/re/encrypt/services/fire/exceptions/MaxRetryOnConnectionReached.java
@@ -1,28 +1,0 @@
-/*
- *
- * Copyright 2018 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-package uk.ac.ebi.ega.file.re.encrypt.services.fire.exceptions;
-
-import java.io.IOException;
-
-public class MaxRetryOnConnectionReached extends IOException {
-
-    public MaxRetryOnConnectionReached() {
-        super("Max retries to connect has been reached");
-    }
-
-}
