+++ fire-core/src/main/java/uk/ac/ebi/ega/fire/IFireFile.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ega.file.re.encrypt.services.fire;
+package uk.ac.ebi.ega.fire;
 
 import java.io.IOException;
 import java.io.InputStream;
